fix(db): ensure PostgreSQL adapter and migration compatibility for Manage Questions

+++ database/migrations/002_ai_generation_one_time_rule.sql
@@ -2,7 +2,7 @@
 -- Safe, idempotent migration script for Online Exam System
 
 ALTER TABLE exams
-ADD COLUMN IF NOT EXISTS ai_generated TINYINT(1) NOT NULL DEFAULT 0;
+ADD COLUMN IF NOT EXISTS ai_generated INT NOT NULL DEFAULT 0;
 
 ALTER TABLE ai_generation_requests
 ADD COLUMN IF NOT EXISTS exam_id INT DEFAULT NULL;
